Db writer impl

+++ modules/tracker/src/main/scala/fi/spectrumlabs/repositories/TrackerCache.scala
@@ -9,20 +9,20 @@ import tofu.higherKind.derived.representableK
 @derive(representableK)
 trait TrackerCache[F[_]] {
 
-  def setLastOffset(offset: Int): F[Unit]
+  def setLastOffset(offset: Long): F[Unit]
 
-  def getLastOffset: F[Int]
+  def getLastOffset: F[Long]
 }
 
 object TrackerCache {
 
-  def create[I[_]: Functor, F[_]: Functor](implicit redis: RedisCommands[F, String, Int]): TrackerCache[F] =
+  def create[I[_]: Functor, F[_]: Functor](implicit redis: RedisCommands[F, String, Long]): TrackerCache[F] =
     new Impl[F]
 
-  private final class Impl[F[_]: Functor](implicit redis: RedisCommands[F, String, Int]) extends TrackerCache[F] {
-    def setLastOffset(offset: Int): F[Unit] = redis.set(Key, offset)
+  private final class Impl[F[_]: Functor](implicit redis: RedisCommands[F, String, Long]) extends TrackerCache[F] {
+    def setLastOffset(offset: Long): F[Unit] = redis.set(Key, offset)
 
-    def getLastOffset: F[Int] = redis.get(Key).map(_.getOrElse(0))
+    def getLastOffset: F[Long] = redis.get(Key).map(_.getOrElse(0))
 
     private val Key: String = "tracker-offset-last"
   }
